✨ feat : 리액트 쿼리 서버사이드 가능하도록 수정
